+++ funnylang/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "1.0.0"
-BYTECODE_VERSION = 1
+__version__ = "1.1.0"
+BYTECODE_VERSION = 2